+++ integration/query_frontend_test.go
@@ -41,7 +41,6 @@ type queryFrontendTestConfig struct {
 	queryStatsEnabled           bool
 	setup                       func(t *testing.T, s *e2e.Scenario) (configFile string, flags map[string]string)
 	withHistograms              bool
-	shardActiveSeriesQueries    bool
 	remoteExecutionEnabled      bool
 }